+++ mac/Keyman4MacIM/Keyman4MacIM/KMInputMethodEventHandler.m
@@ -51,7 +51,7 @@ NSRange _previousSelRange;
     BOOL legacy = ([clientAppId isEqual: @"com.apple.iWork.Pages"] ||
             [clientAppId isEqual: @"com.apple.iWork.Keynote"] ||
             [clientAppId isEqual: @"com.github.atom"] ||
-            [clientAppId isEqual: @"com.collabora.libreoffice-free"] ||
+            [clientAppId containsString: @"libreoffice"] ||
             [clientAppId isEqual: @"com.axosoft.gitkraken"] ||
             [clientAppId isEqual: @"org.sil.app.builder.scripture.ScriptureAppBuilder"] ||
             [clientAppId isEqual: @"org.sil.app.builder.reading.ReadingAppBuilder"] ||