+++ web/testing/platform/utilities.js
@@ -1,0 +1,11 @@
+function loadKeyboards() 
+{ 
+  var kmw=keyman;
+
+  kmw.addKeyboards({id:'platformtest',name:'Platform Testing',
+    languages:{
+      id:'en',name:'English',region:'North America'
+    },
+    filename:'platformtest.js'
+  });
+}