+++ web/testing/test-updateLayer/utilities.js
@@ -1,0 +1,11 @@
+function loadKeyboards() 
+{ 
+  var kmw=keyman;
+
+  kmw.addKeyboards({id:'rac_balti',name:'RAC Balti',
+    languages:{
+      id:'en',name:'English',region:'North America'
+    },
+    filename:'rac_balti.js'
+  });
+}