+++ windows/src/desktop/history.md
@@ -1,5 +1,21 @@
 # Keyman Desktop Version History
 
+## 2018-06-13 10.0.1107.0 beta
+* Add Unicode 11.0 characters
+
+## 2018-06-12 10.0.1106.0 beta
+* No changes in Keyman Desktop
+
+## 2018-06-11 10.0.1105.0 beta
+* Platform tests (in keyboards) are now consistent across all platforms (#969)
+
+## 2018-06-09 10.0.1104.0 beta
+* Keyboards that have script subtags will now install on Windows 7 (#964)
+
+## 2018-06-07 10.0.1103.0 beta
+* Keyman Configuration will now show keyboards that have no name with their filename (#955)
+* Keyboard packages with BCP 47 language subtags unrecognised by Windows will now install on Windows 7 (#948)
+
 ## 2018-06-05 10.0.1102.0 beta
 * Improved crash reporting stability (#940)